fix(event, tests): select_by_name function moved  to conftest
